+++ tests/test_grid.py
@@ -1,4 +1,5 @@
 from matplotlib.path import Path
+from numpy import array, ma
 from pytest import approx
 
 from py_eddy_tracker.data import get_path
@@ -53,3 +54,24 @@ def test_bounds():
     x0, x1, y0, y1 = G.bounds
     assert x0 == -1 / 120.0 and x1 == 360 - 1 / 120
     assert y0 == approx(-90 - 1 / 120.0) and y1 == approx(90 - 1 / 120)
+
+
+def test_interp():
+    # Fake grid
+    g = RegularGridDataset.with_array(
+        coordinates=("x", "y"),
+        datas=dict(
+            z=ma.array(((0, 1), (2, 3)), dtype="f4"),
+            x=array((0, 20)),
+            y=array((0, 10)),
+        ),
+        centered=True,
+    )
+    x0, y0 = array((10,)), array((5,))
+    x1, y1 = array((15,)), array((5,))
+    # Interp nearest
+    assert g.interp("z", x0, y0, method="nearest") == 0
+    assert g.interp("z", x1, y1, method="nearest") == 2
+    # Interp bilinear
+    assert g.interp("z", x0, y0) == 1.5
+    assert g.interp("z", x1, y1) == 2