+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Acces at dataset variable like attribute, and lifetime/age are available for all observations
 - Add **EddyInfos** application to get general information about eddies dataset
 - Add method to inspect contour rejection (which are not in eddies)
+- Grid interp could be "nearest" or "bilinear" 
 ### Changed
 
 ### Removed